requirements.txt
mlx
mlx-lm
argparse
pyyaml
transformers
tqdm